# Conflicts:
#	modules/CMakeLists.txt
#	modules/camera/src/vkcv/camera/Camera.cpp
#	projects/CMakeLists.txt

+++ src/vkcv/Window.cpp
@@ -150,11 +150,15 @@ namespace vkcv {
     }
 
     void Window::onGamepadEvent(int gamepadIndex) {
-        int activeWindowIndex = std::find_if(s_Windows.begin(),
-                                             s_Windows.end(),
-                                             [](GLFWwindow* window){return glfwGetWindowAttrib(window, GLFW_FOCUSED);})
-                                - s_Windows.begin();
-        activeWindowIndex *= (activeWindowIndex < s_Windows.size());    // fixes index getting out of bounds (e.g. if there is no focused window)
+        size_t activeWindowIndex = std::find_if(
+        		s_Windows.begin(),
+        		s_Windows.end(),
+        		[](GLFWwindow* window){return glfwGetWindowAttrib(window, GLFW_FOCUSED);}
+		) - s_Windows.begin();
+	
+		// fixes index getting out of bounds (e.g. if there is no focused window)
+        activeWindowIndex *= (activeWindowIndex < s_Windows.size());
+        
         auto window = static_cast<Window *>(glfwGetWindowUserPointer(s_Windows[activeWindowIndex]));
 
         if (window != nullptr) {
